#refactoring-response-directory
I made a mistake to create the response directory in a wrong path it should be inside the dto directory
