+++ src/main/java/com/personal/projectforum/controller/PostingController.java
@@ -4,8 +4,8 @@ import com.personal.projectforum.domain.constant.FormStatus;
 import com.personal.projectforum.domain.constant.SearchType;
 import com.personal.projectforum.dto.request.PostingRequest;
 import com.personal.projectforum.dto.security.ForumPrincipal;
-import com.personal.projectforum.response.PostingResponse;
-import com.personal.projectforum.response.PostingWithCommentsResponse;
+import com.personal.projectforum.dto.response.PostingResponse;
+import com.personal.projectforum.dto.response.PostingWithCommentsResponse;
 import com.personal.projectforum.service.PaginationService;
 import com.personal.projectforum.service.PostingService;
 import lombok.RequiredArgsConstructor;
