+++ src/main/java/com/personal/projectforum/dto/response/PostingResponse.java
@@ -1,10 +1,8 @@
-package com.personal.projectforum.response;
+package com.personal.projectforum.dto.response;
 
 import com.personal.projectforum.domain.PostingComment;
-import com.personal.projectforum.dto.PostingCommentDto;
 import com.personal.projectforum.dto.PostingDto;
 
-import java.io.Serializable;
 import java.time.LocalDateTime;
 
 /**
