+++ src/main/java/com/personal/projectforum/dto/response/PostingWithCommentsResponse.java
@@ -1,9 +1,8 @@
-package com.personal.projectforum.response;
+package com.personal.projectforum.dto.response;
 
 import com.personal.projectforum.domain.PostingComment;
 import com.personal.projectforum.dto.PostingWithCommentsDto;
 
-import java.io.Serializable;
 import java.time.LocalDateTime;
 import java.util.LinkedHashSet;
 import java.util.Set;
